Refactor utils and correct file imports
