+++ src/components/BotBuilder/BotBuilder.js
@@ -7,9 +7,8 @@ import Snackbar from 'material-ui/Snackbar';
 import Add from 'material-ui/svg-icons/content/add';
 import Delete from 'material-ui/svg-icons/action/delete';
 import { FloatingActionButton, Paper } from 'material-ui';
-import colors from '../../Utils/colors';
 import './BotBuilder.css';
-import urls from '../../Utils/urls';
+import { urls, colors } from '../../utils';
 import { Link } from 'react-router-dom';
 import Cookies from 'universal-cookie';
 import * as $ from 'jquery';
