+++ src/components/BotBuilder/BotWizard.js
@@ -5,7 +5,6 @@ import { Step, Stepper, StepButton } from 'material-ui/Stepper';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
 import { Grid, Col, Row } from 'react-flexbox-grid';
-import colors from '../../Utils/colors';
 import Build from './BotBuilderPages/Build';
 import PropTypes from 'prop-types';
 import Design from './BotBuilderPages/Design';
@@ -18,7 +17,7 @@ import Snackbar from 'material-ui/Snackbar';
 import { Paper, TextField } from 'material-ui';
 import ChevronLeft from 'material-ui/svg-icons/navigation/chevron-left';
 import ChevronRight from 'material-ui/svg-icons/navigation/chevron-right';
-import urls from '../../Utils/urls';
+import { urls, colors } from '../../utils';
 import Icon from 'antd/lib/icon';
 import * as $ from 'jquery';
 import Cookies from 'universal-cookie';
