+++ src/components/SkillEditor/SkillEditor.js
@@ -25,8 +25,7 @@ import * as $ from 'jquery';
 import notification from 'antd/lib/notification';
 import StaticAppBar from '../StaticAppBar/StaticAppBar.react';
 import { red500 } from 'material-ui/styles/colors';
-import colors from '../../Utils/colors';
-import urls from '../../Utils/urls';
+import { urls, colors } from '../../utils';
 const groups = [];
 const languages = [];
 const fontsizes = [];
