+++ src/components/StaticAppBar/StaticAppBar.react.js
@@ -23,9 +23,7 @@ import SKillIcon from 'material-ui/svg-icons/action/dashboard';
 import MenuItem from 'material-ui/MenuItem';
 import { Link } from 'react-router-dom';
 import susiWhite from '../images/SUSIAI-white.png';
-import colors from '../../Utils/colors';
-import urls from '../../Utils/urls';
-import { isProduction } from '../../Utils/helperFunctions';
+import { urls, colors, isProduction } from '../../utils';
 import $ from 'jquery';
 import './StaticAppBar.css';
 // import ListUser from '../Admin/ListUser/ListUser';
