+++ src/utils/index.js
@@ -1,0 +1,7 @@
+export { default as avatars } from './avatars';
+export { default as colors } from './colors';
+export { default as isProduction } from './isProduction';
+export { default as isoCountries } from './isoCountries';
+export { default as parseDate } from './parseDate';
+export { default as request } from './request';
+export { default as urls } from './urls';
