+++ src/utils/isProduction.js
@@ -1,0 +1,6 @@
+let isProduction = () => {
+  let domain = window.location.hostname;
+  return domain.indexOf('.susi.ai') > 0;
+};
+
+export default isProduction;
